Affichage des données calculées

+++ src/mrsrequest/models.py
@@ -195,8 +195,10 @@ class MRSRequest(models.Model):
         return getattr(self, 'STATUS_{}'.format(name.upper()))
 
     def get_taxi_cost(self):
-        return (self.distance * 1.62) + (
-            1.9 * 2 * self.transport_set.count()) * 0.91
+        return Decimal(
+            (self.distance * 1.62)
+            + (1.9 * 2 * self.transport_set.count()) * 0.91
+        )
 
     def get_saving(self):
         if not self.insured_shift:
